memperbaiki data person dan kategori

+++ application/controllers/Person.php
@@ -71,7 +71,7 @@ class Person extends CI_Controller {
         $config ['max_size'] = '5500';
         $this->upload->initialize($config); // meng set config yang sudah di atur
 
-        if ( ($this->input->post('nama') != NULL) and ($this->input->post('deskripsi') != NULL) ) {
+        if ( ($this->input->post('nama') != NULL) and ($this->input->post('deskripsi_EN') != NULL) and ($this->input->post('deskripsi_ID') != NULL) ) {
             if (isset($file_image['name']) and $file_image['name'] != '') {
                 if (!$this->upload->do_upload('foto')) {
                     $alert = $this->upload->display_errors();
@@ -80,11 +80,11 @@ class Person extends CI_Controller {
                     $data = array(
                         'name' => $this->input->post('nama'),
                         'jabatan' => $this->input->post('jabatan'),
-                        'deskripsi' => $this->input->post('deskripsi'),
+                        'deskripsi_EN' => $this->input->post('deskripsi_EN'),
+                        'deskripsi_ID' => $this->input->post('deskripsi_ID'),
                         'create_date'=> date('Y-m-d h:i:s'),
                         'image'=> $file_image['name'],
-                        'idcategory'=> $this->input->post('idcategory'),
-                        'idsubcategory'=> $this->input->post('idsubcategory')
+                        'idcategory'=> $this->input->post('idcategory')
                     );
                     if($this->model_person->save_person($data)){
                         $alert = url_title("Saved!");
@@ -113,16 +113,16 @@ class Person extends CI_Controller {
                 'Dashboard'=>base_url().'admin',
                 'Halaman list' => base_url().'Halaman'
                 ),
-            'title'=> 'Halaman <small>management</small>',
+            'title'=> 'Employee <small>management</small>',
             'last_login' => $this->sess['last_login'],
             'session' => $this->sess['username'],
             'site_lang'=>$this->session->userdata('site_lang'),
             );
         
-        if (isset($_GET['n']) and $_GET['n'] != NULL){
+        if (isset($_GET['n']) and $_GET['n'] != NULL and $_GET['n'] == TRUE){
             $this->model_person->delete($id);
             $alert = url_title("delete succses !");
-            redirect('page/?n='.$alert,'refresh');   
+            redirect('person/?n='.$alert,'refresh');   
         }
         
         $this->smartyci->assign('data',$data);
@@ -154,26 +154,62 @@ class Person extends CI_Controller {
     }
 
     public function update(){
-        $data = array(
-          'title_EN' => $this->input->post('title_EN'),
-                'title_ID' => $this->input->post('title_ID'),
-                'keyword_EN' => $this->input->post('title_EN'),
-                'keyword_ID' => $this->input->post('title_EN'),
-                'content_EN' => $this->input->post('content_EN'),
-                'content_ID' => $this->input->post('content_ID'),
-                'waktu'=> date('Y-m-d h:i:s'),
-                'sidebar'=> $this->input->post('sidebar'),
-                'url'=> $this->input->post('url')
-        );
 
-        $id = $this->input->post('id');
-        //print_r($data);
-        if($this->model_person->update_page($id,$data)) {
-            $alert = url_title("Update succses !");
-            redirect('page/edit/'.$id.'?n='.$alert,'refresh');
-        }else{
-            $alert = url_title("Save failde !");
-            redirect('page/edit/'.$id.'?n='.$alert,'refresh');
+        if ($_FILES['foto'] != NULL) 
+        {
+            $this->load->library('upload');
+            $file_image = (isset($_FILES['foto']) == TRUE ? $_FILES['foto'] : null); // ambil dahulu
+            $config ['upload_path'] = "./document/images/employee/"; // lokasi folder yang akan digunakan untuk menyimpan file
+            $config ['allowed_types'] = 'JPG|jpg|JPEG|jpeg|PNG|png'; // extension yang diperbolehkan untuk diupload
+            // $config ['file_name'] = url_title(slug($this->input->post('nama')))."-employee";
+            $config ['max_size'] = '5500';
+            $this->upload->initialize($config); // meng set config yang sudah di atur
+
+            if (isset($file_image['name']) and $file_image['name'] != '') {
+                if (!$this->upload->do_upload('foto')) {
+                    $alert = $this->upload->display_errors();
+                    redirect('person/add?n=' . $alert, 'refresh');
+                } else {
+
+                    $data = array(
+                        'name' => $this->input->post('nama'),
+                        'jabatan' => $this->input->post('jabatan'),
+                        'deskripsi_EN' => $this->input->post('deskripsi_EN'),
+                        'deskripsi_ID' => $this->input->post('deskripsi_ID'),
+                        'create_date'=> date('Y-m-d h:i:s'),
+                        'image'=> $file_image['name'],
+                        'idcategory'=> $this->input->post('idcategory')
+                    );
+
+                    $id = $this->input->post('id');
+                    if($this->model_person->update($id,$data)) {
+                        $alert = url_title("Update succses !");
+                        redirect('person/edit/'.$id.'?n='.$alert,'refresh');
+                    }else{
+                        $alert = url_title("Save failde !");
+                        redirect('person/edit/'.$id.'?n='.$alert,'refresh');
+                    }
+                }
+        } else {
+            $data = array(
+                'name' => $this->input->post('nama'),
+                'jabatan' => $this->input->post('jabatan'),
+                'deskripsi_EN' => $this->input->post('deskripsi_EN'),
+                'deskripsi_ID' => $this->input->post('deskripsi_ID'),
+                'create_date'=> date('Y-m-d h:i:s'),
+                'image'=> $file_image['name'],
+                'idcategory'=> $this->input->post('idcategory')
+            );
+
+            $id = $this->input->post('id');
+            if($this->model_person->update($id,$data)) {
+                $alert = url_title("Update succses !");
+                redirect('person/edit/'.$id.'?n='.$alert,'refresh');
+            }else{
+                $alert = url_title("Save failde !");
+                redirect('person/edit/'.$id.'?n='.$alert,'refresh');
+            }
         }
+    }
     }
 }
